Refactor: Create project/Lib.scala
